+++ src/main/java/sagex/phoenix/menu/IMenuDelegates.java
@@ -1,0 +1,9 @@
+package sagex.phoenix.menu;
+
+/**
+ * Created by seans on 30/12/16.
+ */
+public interface IMenuDelegates {
+    IMenuItem getOriginalItem();
+    IMenuItem getDelegateItem();
+}